+++ lib/utils/app_utils.dart
@@ -5,9 +5,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:notekar/models/palette.dart';
 
-const appVersion = '7.3.2';
+const appVersion = '7.3.3';
 const kAppVersion = appVersion;
-const kAppBuildNumber = '26BR0906a';
+const kAppBuildNumber = '26BR0906b';
 const appBuildDate = '2026-09-06';
 
 String formatInstalledDate(String buildDateStr) {